backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-socketio>=5.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.24.0
